redesign trace replay cli
